refactor(dockview-modules): move LayoutHistory impl to the modules package (Phase E)
Relocate `LayoutHistoryService` + `LayoutHistoryModule` from `dockview-core`
into `dockview-modules`, the same path the other extracted modules took. Core
keeps the contracts and wiring; nothing user-facing changes.

- Contracts to core `moduleContracts.ts`: `ILayoutHistoryHost`,
  `ILayoutHistoryService`, `LayoutHistoryChangeEvent`, `LayoutHistoryKind`
  (exported from the core index). Option types stay in `options.ts`. The
  component keeps its `?.`-delegating undo/redo/canUndo/canRedo/clearHistory/
  onDidChangeHistory surface + a local never-event fallback.
- Implementation moves to `dockview-modules/src/layoutHistoryService.ts`,
  added to `Modules` (auto-registered on import) and exported. Removed from
  core `allModules`.
- Unit tests move to `dockview-modules/src/__tests__/layoutHistory.spec.ts`
  (the module is registered there via the test setup).
- Regenerated `__generated__/dockview-core-exports.txt` (gen:check).

Removability holds: with the module absent, `api.undo/redo` no-op, `canUndo/
canRedo` are false, `onDidChangeHistory` is the never-event.

Test: 1080 core + 125 modules + 8 e2e green; lint 0 errors; gen clean.
